Got many-to-many assoication to work between exercises table and objectives table, had to do away with one-to-many association to between objectives table and teachers table to make the many-to-many assocaition work
