+++ fixtures/actors.json
@@ -2,58 +2,19 @@
     {
         "model": "actors",
         "data": {
-            "exercise": 1,
-            "book_title": "42 Etudes for Violin",
-            "author_composer": "Kreutzer, Rodolphe",
-            "primary_positions": "1",
-            "secondary_positions": "2, 3, 4",
-            "primary_bowing": "legato",
-            "secondary_bowing": null,
-            "musical_key": "A Minor",
-            "difficulty": "Medium",
-            "focus": "Tone",
-            "type": "Etude",
-            "image": null,
-            "link": null,
-            "book_description": null
+            "actorName": "Tom Hanks"
         }
     },
     {
         "model": "actors",
         "data": {
-            "exercise": 2,
-            "book_title": "42 Etudes for Violin",
-            "author_composer": "Kreutzer, Rodolphe",
-            "primary_positions": "1",
-            "secondary_positions": "2, 3, 4",
-            "primary_bowing": "legato",
-            "secondary_bowing": null,
-            "musical_key": "C Major",
-            "difficulty": "Medium",
-            "focus": "Facility",
-            "type": "Etude",
-            "image": null,
-            "link": null,
-            "book_description": null
+            "actorName": "Gary Siense"
         }
     },
     {
         "model": "actors",
         "data": {
-            "exercise": 3,
-            "book_title": "42 Etudes for Violin",
-            "author_composer": "Kreutzer, Rodolphe",
-            "primary_positions": "1",
-            "secondary_positions": "2, 3, 4, 5, 5+",
-            "primary_bowing": "detache",
-            "secondary_bowing": null,
-            "musical_key": "C Major",
-            "difficulty": "Medium",
-            "focus": "Facility",
-            "type": "Etude",
-            "image": null,
-            "link": null,
-            "book_description": null
+            "actorName": "Bill Paxton"
         }
     }
 ]
