+++ fixtures/movies.json
@@ -2,25 +2,25 @@
     {
         "model": "movies",
         "data": {
-            "objective": "Our first lesson plan",
-            "lesson_plan": "Here is the description for this lesson plan",
-            "actors": [1,2]
+            "movieName": "Forrest Gump",
+            "actors": [1,2],
+            "productionId": 1
         }
     },
     {
         "model": "movies",
         "data": {
-            "objective": "Another lesson plan",
-            "lesson_plan": "Play violin",
-            "actors": [1,3]
+            "movieName": "Apollo 13",
+            "actors": [1,2,3],
+            "productionId": 2
         }
     },
     {
         "model": "movies",
         "data": {
-            "objective": "Music to my ears",
-            "lesson_plan": "Maybe this one is harder",
-            "actors": [2,3]
+            "movieName": "Tombstone",
+            "actors": [3],
+            "productionId": 3
         }
     }
 ]
