+++ fixtures/objectives.json
@@ -2,10 +2,25 @@
     {
         "model": "Objectives",
         "data": {
-            "TeacherId": 1,
             "objective": "Our first lesson plan",
             "lesson_plan": "Here is the description for this lesson plan",
-            "Exercises": [1,2,3]
+            "Exercises": [1,2]
+        }
+    },
+    {
+        "model": "Objectives",
+        "data": {
+            "objective": "Next Lesson",
+            "lesson_plan": "Step this up",
+            "Exercises": [1,3]
+        }
+    },
+    {
+        "model": "Objectives",
+        "data": {
+            "objective": "This one is harder",
+            "lesson_plan": "Bring your A game",
+            "Exercises": [2,3]
         }
     }
 ]
